refactor(site): move chatProviderConfigs to aiProviders module
