--- site/src/api/queries/aiProviderKeys.ts
@@ -1,1 +1,0 @@
-export const aiProvidersKey = ["ai", "providers"] as const;
