+++ site/src/api/queries/aiProviders.ts
@@ -1,20 +1,48 @@
-import type { QueryClient } from "react-query";
+import { type QueryClient, queryOptions } from "react-query";
 import { API } from "#/api/api";
-import { aiProvidersKey } from "#/api/queries/aiProviderKeys";
 import { invalidateChatProviderDependentQueries } from "#/api/queries/chats";
 import type {
 	AIProvider,
+	ChatProviderConfig,
 	CreateAIProviderRequest,
 	UpdateAIProviderRequest,
 } from "#/api/typesGenerated";
 
+const aiProvidersListKey = ["ai", "providers"] as const;
+
 export const aiProviderKeyFor = (idOrName: string) =>
-	[...aiProvidersKey, idOrName] as const;
+	[...aiProvidersListKey, idOrName] as const;
 
 export const aiProvidersList = () => ({
-	queryKey: aiProvidersKey,
+	queryKey: aiProvidersListKey,
 	queryFn: (): Promise<AIProvider[]> => API.getAIProviders(),
 });
+
+const selectChatProviderConfigs = (
+	providers: readonly AIProvider[],
+): ChatProviderConfig[] =>
+	providers.map((provider) => ({
+		id: provider.id,
+		provider: provider.type,
+		display_name: provider.display_name || provider.type,
+		icon: provider.icon,
+		enabled: provider.enabled,
+		has_api_key: provider.api_keys.length > 0,
+		central_api_key_enabled: true,
+		allow_user_api_key: true,
+		allow_central_api_key_fallback: true,
+		base_url: provider.base_url,
+		source: "database",
+		created_at: provider.created_at,
+		updated_at: provider.updated_at,
+	}));
+
+export const chatProviderConfigs = () =>
+	queryOptions({
+		queryKey: aiProvidersListKey,
+		queryFn: (): Promise<AIProvider[]> => API.getAIProviders(),
+		select: selectChatProviderConfigs,
+	});
 
 export const aiProvider = (idOrName: string) => ({
 	queryKey: aiProviderKeyFor(idOrName),
@@ -26,7 +54,7 @@ export const createAIProviderMutation = (queryClient: QueryClient) => ({
 		API.createAIProvider(request),
 	onSuccess: async () => {
 		await Promise.all([
-			queryClient.invalidateQueries({ queryKey: aiProvidersKey }),
+			queryClient.invalidateQueries({ queryKey: aiProvidersListKey }),
 			invalidateChatProviderDependentQueries(queryClient),
 		]);
 	},
@@ -40,7 +68,7 @@ export const updateAIProviderMutation = (
 		API.updateAIProvider(idOrName, request),
 	onSuccess: async () => {
 		await Promise.all([
-			queryClient.invalidateQueries({ queryKey: aiProvidersKey }),
+			queryClient.invalidateQueries({ queryKey: aiProvidersListKey }),
 			queryClient.invalidateQueries({
 				queryKey: aiProviderKeyFor(idOrName),
 			}),
@@ -57,7 +85,7 @@ export const deleteAIProviderMutation = (
 	onSuccess: async () => {
 		queryClient.removeQueries({ queryKey: aiProviderKeyFor(idOrName) });
 		await Promise.all([
-			queryClient.invalidateQueries({ queryKey: aiProvidersKey }),
+			queryClient.invalidateQueries({ queryKey: aiProvidersListKey }),
 			invalidateChatProviderDependentQueries(queryClient),
 		]);
 	},
